requirements.txt
Django==1.7
Markdown==2.5
Pillow==2.5.3
argparse==1.2.1
djangorestframework==2.4.3
wsgiref==0.1.2
